Trait streebog_hash::StreebogHasher

pub trait StreebogHasher {
    fn new() -> Self;

    fn update(&mut self, data_chunk: &[u8]);

    fn finish(&mut self);

    fn get_result(&self) -> Box<[u8]>;

    fn get_result_str(&self) -> String;

    fn reset(&mut self);
}

A trait which represents the ability to hash an arbitrary stream of bytes with Streebog algorithm.

Required Methods

fn new() -> Self

Creates new hasher object.

fn update(&mut self, data_chunk: &[u8])

Writes some data into this hasher.

fn finish(&mut self)

Completes a round of hashing.

fn get_result(&self) -> Box<[u8]>

Returns result of hashing as Box<[u8]> (result is big-endian, i.e. bytes arranged in the same order as in String representation).

If hasher is not finished (i.e. finish is not called), it returns empty array.

fn get_result_str(&self) -> String

Returns result of hashing as String.

If hasher is not finished (i.e. finish is not called), it returns empty String.

fn reset(&mut self)

Reset hasher to default state and mark as not finished.

Note! After calling this all processed data will be lost!